+++ extras/PackageMaintenance.R
@@ -79,6 +79,7 @@ rCode <- ParallelLogger::createArgFunction("trimByPsToEquipoise",
                                            excludeArgs = c("population"),
                                            rCode = rCode)
 rCode <- ParallelLogger::createArgFunction("trimByIptw", excludeArgs = c("population"), rCode = rCode)
+rCode <- ParallelLogger::createArgFunction("truncateIptw", excludeArgs = c("population"), rCode = rCode)
 rCode <- ParallelLogger::createArgFunction("matchOnPs", excludeArgs = c("population"), rCode = rCode)
 rCode <- ParallelLogger::createArgFunction("matchOnPsAndCovariates",
                                            excludeArgs = c("population", "cohortMethodData"),